requirements.txt
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
djangorestframework==3.11.0
idna==2.9
psycopg2==2.8.4
psycopg2-binary==2.8.4
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.8
